backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow==11.1.0
pytesseract==0.3.13
invoice2data==0.4.5
anthropic==0.43.0
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2
sse-starlette==2.1.3
sqlalchemy==2.0.36
alembic==1.14.1
PyJWT==2.10.1
cryptography
python-dateutil
